+++ sandbox/urls.py
@@ -1,13 +1,9 @@
+from django.apps import apps
 from django.conf import settings
 from django.conf.urls import include, url
 from django.contrib import admin
 from django.conf.urls import i18n as i18n_urls
 from django.views.static import serve
-from oscar.app import application as oscar_application
-from oscarapi.app import application as oscar_api
-from oscarapicheckout.app import application as oscar_api_checkout
-from wellsfargo.api.app import application as wfrs_api
-from wellsfargo.dashboard.app import application as wfrs_app
 
 
 urlpatterns = [
@@ -16,11 +12,11 @@ urlpatterns = [
     url(r'^media/(?P<path>.*)$', serve, {'document_root': settings.MEDIA_ROOT, 'show_indexes': True}),
 
     # Include plugins
-    url(r'^dashboard/wfrs/', wfrs_app.urls),
-    url(r'^api/wfrs/', wfrs_api.urls),
-    url(r'^api/', oscar_api_checkout.urls),
-    url(r'^api/', oscar_api.urls),
+    url(r'^dashboard/wfrs/', include(apps.get_app_config('wellsfargo_dashboard').urls[0])),
+    url(r'^api/wfrs/', include(apps.get_app_config('wellsfargo_api').urls[0])),
+    url(r'^api/', include(apps.get_app_config('oscarapicheckout').urls[0])),
+    url(r'^api/', include('oscarapi.urls')),
 
     # Include stock Oscar
-    url(r'', oscar_application.urls),
+    url(r'^', include(apps.get_app_config('oscar').urls[0])),
 ]