+++ src/wellsfargo/__init__.py
@@ -1,5 +1,1 @@
-import os.path
-
-default_app_config = 'wellsfargo.config.Config'
-
-WFRS_TEMPLATE_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'templates')
+default_app_config = 'wellsfargo.apps.WFRSConfig'