+++ src/wellsfargo/apps.py
@@ -1,12 +1,13 @@
-from django.apps import AppConfig
 from django.utils.translation import gettext_lazy as _
+from oscar.core.application import OscarConfig
 
 
-class Config(AppConfig):
+class WFRSConfig(OscarConfig):
     name = 'wellsfargo'
     label = 'wellsfargo'
     # Translators: Backend Library Name
     verbose_name = _('Wells Fargo Retail Services')
+    namespace = 'wellsfargo'
 
     def ready(self):
         from . import handlers  # NOQA