+++ src/wellsfargo/tests/api/test_api_credit_app.py
@@ -3,7 +3,7 @@ from rest_framework.reverse import reverse
 from wellsfargo.models import USCreditApp
 from wellsfargo.tests.base import BaseTest
 from wellsfargo.tests import responses
-import mock
+from unittest import mock
 
 
 class CreditApplicationSelectorTest(BaseTest):